Add SRV3 decoder
Preliminary support for decoding the SRV3 subtitle format.
SRV3 is the internal format YouTube uses for their captions. Supporting it
in ffmpeg allows video players to play a significant subset of SRV3
mostly correctly by converting it to ASS.
Currently the following features are unsupported:
- Vertical text
- Scrolling text
- Ruby text
- Background box support is incomplete
These issues are mostly due to limitations of the ASS format.

Signed-off-by: Hubert Głuchowski <fishhh@fishhh.dev>